🚀 Add About Us page and navigation
This commit adds the About Us page to the app. It includes a map with a marker indicating the location of EPSI. The user can click on the floating action button to center the map on EPSI's location. Additionally, a new route named 'aboutus' is added to navigate to the About Us page from the Home page using an IconButton with an info icon.

Dependencies:
- Added flutter_map version 6.1.0
- Added latlong2 version 0.9.0